requirements.txt
dotenv
google-generativeai
Pillow